requirements.txt
requests
pandas
lxml
openai
python-dotenv
beautifulsoup4
tqdm
PyMySQL
langchain
langchain-community
llama-index